% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_git_remotes.R
\name{add_git_remotes}
\alias{add_git_remotes}
\title{add_git_remotes}
\usage{
add_git_remotes(force_bb = FALSE)
}
\arguments{
\item{force_bb}{Bitbucket repos can not be directly queried; this option
simply forces a Bitbucket remote to be added regardless of whether or not
that remote repository actually exists or not.}
}
\description{
Automatically add git remotes for Gitlab, Bitbucket, and Sourcehut if the
remote repositories exist.
}
