% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug}
\alias{debug}
\title{Configure an R package path for debugging with 'gdb' or 'lldb'}
\usage{
debug()
}
\value{
Nothing; this just modifies or writes a 'Makevars' script. The
debugger can then be started with a single extra command.
}
\description{
Note that no path arguments are used or recognised, and this function can
only be called and used in the current working directory.
}
