\name{dfprint}
\alias{dfprint}
\title{Print a data.frame to HTML.}
\usage{
  dfprint(df, printdigits = 2)
}
\arguments{
  \item{df}{data.frame to print}

  \item{printdigits}{the number of digits to print for
  numeric variables}
}
\description{
  This function simplifies printing data.frames to HTML.
}
\examples{
dfprint(Orange)
}
\keyword{prettyprint}

