% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdb.R
\name{gdb}
\alias{gdb}
\title{Configure an R package path for debugging with 'gdb'}
\usage{
gdb()
}
\value{
Nothing; calling this will start the 'gdb' debugger on the nominated
script.
}
\description{
Note that no path arguments are used or recognised, and this function can
only be called and used in the current working directory.
}
