% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-commits.R
\name{gh_daily_commits}
\alias{gh_daily_commits}
\title{Count GitHub commits for current day}
\usage{
gh_daily_commits(quiet = FALSE)
}
\arguments{
\item{quiet}{If \code{FALSE}, print notifications to screen}
}
\value{
Single integer count of daily commits
}
\description{
Count GitHub commits for current day
}
