% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_version.R
\name{increment_dev_version}
\alias{increment_dev_version}
\title{increment_dev_version}
\usage{
increment_dev_version()
}
\description{
Increment development version number in DESCRIPTION and 'codemeta.json' if it
exists. Current directory must be within an R package.
}
\details{
The 'usethis::use_dev_version()' function only adds a fixed '.9000' as the
fourth component. This function is intended to be used to incremenet this
fourth component on every commit. It will also update the version number in
'codemeta.json' as generated by the 'codemetar' package, if that file exists.

This function can easily be alised for ease of use, for example in
'~/.bash_aliases' as 'incr="Rscript -e 'mpmisc::increment_dev_version()'"'.
You then only need to type 'incr' in your shell to automatically increment
the version.
}
