% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-notifications.R
\name{mark_gh_notifications_as_read}
\alias{mark_gh_notifications_as_read}
\title{Mark all new GitHub notifications as read and remove from list}
\usage{
mark_gh_notifications_as_read()
}
\description{
This does not remove them from the top of the most recent notifications list,
but just marks them as read. There is currently no way to remove them from
the list.
}
