\docType{package}
\name{mpmisc-package}
\alias{mpmisc}
\alias{mpmisc-package}
\title{mpmisc: a collection of miscellaneous functions for report-making.}
\description{
  This package provides a collection of miscellaneous
  functions for report-making around the TB Clinic.
}
\author{
  Matt Parker \email{mattmparker@gmail.com}
}
\keyword{misc}
\keyword{package}

