% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{mpmisc_daily}
\alias{mpmisc_daily}
\title{Add text of daily activity to 'daily.md' file}
\usage{
mpmisc_daily(txt = "")
}
\arguments{
\item{txt}{A string summarising daily activity.}
}
\description{
Add text of daily activity to 'daily.md' file
}
