% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-issues.R
\name{open_gh_issue}
\alias{open_gh_issue}
\title{Open web browser to specified GitHub issue}
\usage{
open_gh_issue(repo = NULL, issue = NULL, user = "mpadge")
}
\arguments{
\item{repo}{Name of repo}

\item{issue}{Optional number of issue}
}
\value{
Nothing
}
\description{
Open issue only with name of repo, or if no issue specified open main issues
page of repo. This uses \link{latest_comments} to extract issues for
specified user in order to associate repo name with the correct GitHub
organization.
}
