% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-notifications.R
\name{open_gh_notification}
\alias{open_gh_notification}
\title{Open web browser to specified GitHub notification}
\usage{
open_gh_notification(what = "github", n)
}
\arguments{
\item{what}{Service to get notifications from; one of "github" or
"codeberg" (case-insensitive).}

\item{n}{Number of notification as generated from
\link{gh_notifications}}
}
\value{
Nothing
}
\description{
This accepts only the single parameter which corresponds to a number from the
most recently extracted and cached notification list.
}
