% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixabay.R
\name{pixabay}
\alias{pixabay}
\title{Grab random pixabay image for leftwm background}
\usage{
pixabay(category = "buildings")
}
\arguments{
\item{category}{Pixabay category from list at
\url{https://pixabay.com/api/docs/#api_search_images}.}
}
\description{
Grab random pixabay image for leftwm background
}
