% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-univ-status.R
\name{r_univ_build_status}
\alias{r_univ_build_status}
\title{Function to check status on R-Universe and issue any warnings.}
\usage{
r_univ_build_status(univ = "ropensci-review-tools", maintainer = NULL)
}
\arguments{
\item{univ}{Name of universe, as sub-domain prefix of main R-universe URL.}

\item{maintainer}{Optional name of maintainer. If given, results with be
filtered to only those packages within specified univsere maintained by that
person.}
}
\value{
Nothing. Function is called only for its side-effect of generating
messages in an active terminal session.
}
\description{
Function to check status on R-Universe and issue any warnings.
}
