% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-updates.R
\name{workflow_badges}
\alias{workflow_badges}
\title{Identify any README.md badge links to workflows with names different from
actual workflow values.}
\usage{
workflow_badges(root = ".", max_depth = 2L)
}
\arguments{
\item{root}{Directory in which function is to be run.}

\item{max_depth}{Maximal recursive depth for trawling sub-directories.}
}
\description{
Scans all directories above where function is called.
}
