% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-updates.R
\name{workflow_updates}
\alias{workflow_updates}
\title{A local-only version of dependabot.}
\usage{
workflow_updates(root = ".", max_depth = 2L)
}
\arguments{
\item{root}{Directory in which function is to be run.}

\item{max_depth}{Maximal recursive depth for trawling sub-directories.}
}
\value{
(Invisibly) A \pkg{tibble} \code{data.frame} of all obsolete workflow
versions. Function is primarily intended to be called for the side-effect of
printing all results to screen.
}
\description{
Scans all directories above where function is called, extracts maximum
version number of all unique GitHub actions used in every workflow, and
reports paths to any repositories which use workflow versions less than
those.
}
