% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-notifications.R
\name{gh_notifications}
\alias{gh_notifications}
\title{Get latest unread notifications from GitHub or Codeberg}
\usage{
gh_notifications(where = "github", quiet = FALSE)
}
\arguments{
\item{where}{Service to get notifications from; one of "github" or
"codeberg" (case-insensitive).}

\item{quiet}{If \code{FALSE}, print notifications to screen}
}
\value{
\code{data.frame} of notifications (invisibly)
}
\description{
Not yet on graphql API:
https://github.community/t/get-notification-list-via-graphql-api/13836
but available on REST API:
https://docs.github.com/en/rest/reference/activity#notifications
And Codeberg API:
https://codeberg.org/api/swagger#/notification
}
