% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-issues.R
\name{gh_notifications}
\alias{gh_notifications}
\title{Get latest unread github notifications}
\usage{
gh_notifications(quiet = FALSE)
}
\arguments{
\item{quiet}{If \code{FALSE}, print notifications to screen}
}
\value{
\code{data.frame} of notifications (invisibly)
}
\description{
Get latest unread github notifications
}
