% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-notifications.R
\name{mark_gh_notifications_as_read}
\alias{mark_gh_notifications_as_read}
\title{Mark all new GitHub notifications as read and remove from list}
\usage{
mark_gh_notifications_as_read()
}
\description{
Mark all new GitHub notifications as read and remove from list
}
