% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpmisc-package.R
\docType{package}
\name{mpmisc-package}
\alias{mpmisc}
\alias{mpmisc-package}
\title{mpmisc: Miscellaneous functions which I load in my '.Rprofile'}
\description{
Miscellaneous functions which I load in my '.Rprofile'.
}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
